created seperate file containing all logic to get recipe to display from params
