+++ src/pages/RecipeView/index.tsx
@@ -1,26 +1,10 @@
-import { useParams } from "react-router-dom";
-import { useContext } from "react";
-import { RecipeContextType } from "../../shared/types/RecipeContextType.ts";
-import { RecipeContext } from "../../shared/components/RecipeContextProvider";
-import { Recipe } from "../../shared/types/Recipe.ts";
 import { RecipeCard } from "../../components/RecipeCard";
 import { PageWidthContainer } from "../../shared/components/PageWidthContainer.tsx";
 import { StyledRecipeBody } from "./RecipeView.styled.tsx";
-
-type ParamsInterface = {
-  recipeId: string;
-};
+import { getRecipeToDisplayFromParams } from "../../shared/getRecipeToDisplayFromParams.tsx";
 
 export function RecipeView() {
-  const { savedRecipes } = useContext<RecipeContextType>(RecipeContext);
-  const params = useParams<ParamsInterface>();
-  let recipeId = null;
-  if (params.recipeId) {
-    recipeId = Number(params.recipeId);
-  }
-  const recipeToDisplay = savedRecipes?.find(
-    (recipe: Recipe) => recipe.id === recipeId,
-  );
+  const recipeToDisplay = getRecipeToDisplayFromParams();
 
   return (
     <PageWidthContainer>
